Create captcha_form_builder and add it's methods to FormBuilder.
