+++ lib/yacaph.rb
@@ -6,7 +6,7 @@ require 'yacaph/engine' if Rails::VERSION::MAJOR >= 3
 require "yacaph/captcha_util"
 require "yacaph/captcha_helper"
 require "yacaph/validates_captcha"
-require "yacaph/form_builder"
+require "yacaph/captcha_form_builder"
 
 AbstractController::Base.class_eval do
   include ValidatesCaptcha
@@ -14,4 +14,8 @@ end
 
 ActiveModel::Validations.class_eval do
   include ValidatesCaptcha
+end
+
+ActionView::Helpers::FormBuilder.class_eval do
+  include CaptchaFormBuilder
 end
